update number display
